create ParameterInfoUtil helper class for junit tests GP-5519
