+++ test/src/org/genepattern/server/job/input/choice/TestChoiceInfo.java
@@ -3,8 +3,8 @@
  *******************************************************************************/
 package org.genepattern.server.job.input.choice;
 
-import java.util.HashMap;
 
+import org.genepattern.junitutil.ParameterInfoUtil;
 import org.genepattern.webservice.ParameterInfo;
 import org.junit.Assert;
 import org.junit.Test;
@@ -16,57 +16,19 @@ import org.junit.Test;
  */    
 @SuppressWarnings("unchecked")
 public class TestChoiceInfo {
+    public static final String pname="input.file";
     public static final String dropDownDir="ftp://gpftp.broadinstitute.org/example_data/gpservertest/DemoFileDropdown/input.dir/";
-    
-    /**
-     * For junit testing, initialize a ParameterInfo which has a choiceDir.
-     * @param choiceDir
-     * @return
-     */
-    public static final ParameterInfo initFtpParam(final String choiceDir) {
-        final String name="input.file";
-        final String value="";
-        final String description="A file drop-down";
-    
-        final ParameterInfo pinfo=new ParameterInfo(name, value, description);
-        pinfo.setAttributes(new HashMap<String,String>());
-        pinfo.getAttributes().put("MODE", "IN");
-        pinfo.getAttributes().put("TYPE", "FILE");
-        pinfo.getAttributes().put("choiceDir", choiceDir);
-        pinfo.getAttributes().put("default_value", "");
-        pinfo.getAttributes().put("fileFormat", "txt");
-        pinfo.getAttributes().put("flag", "");
-        pinfo.getAttributes().put("optional", "");
-        pinfo.getAttributes().put("prefix", "");
-        pinfo.getAttributes().put("prefix_when_specified", "");
-        pinfo.getAttributes().put("type", "java.io.File");
-    
-        return pinfo;
-    }
-    
-    public static final ParameterInfo initFileParam(String name, String value, String description) {
-        final ParameterInfo pinfo=new ParameterInfo(name, value, description);
-        pinfo.setAttributes(new HashMap<String,String>());        
-        pinfo.getAttributes().put("MODE", "IN");
-        pinfo.getAttributes().put("TYPE", "FILE");
-        pinfo.getAttributes().put("default_value", "");
-        pinfo.getAttributes().put("optional", "");
-        pinfo.getAttributes().put("prefix_when_specified", "");
-        pinfo.getAttributes().put("type", "java.io.File");
-
-        return pinfo;
-    } 
     
     @Test
     public void testFtpPassiveMode_default() {
-        final ParameterInfo pinfo=initFtpParam(dropDownDir);
+        final ParameterInfo pinfo=ParameterInfoUtil.initFileDropdownParam(pname, dropDownDir);
         Assert.assertEquals("by default, passiveMode is true", true, 
                 ChoiceInfo.getFtpPassiveMode(pinfo));
     }
     
     @Test
     public void testFtpPassiveMode_declareTrue() {
-        final ParameterInfo pinfo=initFtpParam(dropDownDir);
+        final ParameterInfo pinfo=ParameterInfoUtil.initFileDropdownParam(pname, dropDownDir);
         pinfo.getAttributes().put(ChoiceInfo.PROP_CHOICE_DIR_FTP_PASSIVE_MODE, "true");
         Assert.assertEquals("passiveMode=true", true, 
                 ChoiceInfo.getFtpPassiveMode(pinfo));
@@ -74,7 +36,7 @@ public class TestChoiceInfo {
 
     @Test
     public void testFtpPassiveMode_declareFalse() {
-        final ParameterInfo pinfo=initFtpParam(dropDownDir);
+        final ParameterInfo pinfo=ParameterInfoUtil.initFileDropdownParam(pname, dropDownDir);
         pinfo.getAttributes().put(ChoiceInfo.PROP_CHOICE_DIR_FTP_PASSIVE_MODE, "false");
         Assert.assertEquals("passiveMode=false", false, 
                 ChoiceInfo.getFtpPassiveMode(pinfo));
@@ -82,7 +44,7 @@ public class TestChoiceInfo {
 
     @Test
     public void testFtpPassiveMode_caseInsensitive() {
-        final ParameterInfo pinfo=initFtpParam(dropDownDir);
+        final ParameterInfo pinfo=ParameterInfoUtil.initFileDropdownParam(pname, dropDownDir);
         pinfo.getAttributes().put(ChoiceInfo.PROP_CHOICE_DIR_FTP_PASSIVE_MODE, "FALSE");
         Assert.assertEquals("passiveMode=false", false, 
                 ChoiceInfo.getFtpPassiveMode(pinfo));
